tdf#153918 svx a11y: Drop old items when switching FontCharMap
Calling `rRenderContext.GetFontCharMap(mxFontCharMap)`
in `SvxShowCharSet::RecalculateFont` means that a new
`FontCharMapRef` gets assigned to `mxFontCharMap`.

Therefore, clear the old items based on the previous
one from the map, so that new ones will be created
based on the new font char map in `SvxShowCharSet::ImplGetItem`
as needed instead of still returning the old ones.

Adapt the a11y UI test to check again that the "!" character
now has the proper accessible name. The incorrect
accessible name in the special characters dialog had
actually been uncovered by the initial version of the
test that failed on Windows due to this, and the test was adapted
to workaround that, s. tdf#153918 and the discussion in the
Gerrit change adding the test [1] for more details.

[1] https://gerrit.libreoffice.org/c/core/+/142260/24..26/sw/qa/extras/accessibility/dialogs.cxx#b85

Change-Id: I0ac578750b1fa0a7e63e62a6937a125c7ceab510
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148423
Reviewed-by: Colomban Wendling <cwendling@hypra.fr>
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>