git-hooks: better handling in './g -z' with win-links
  and linux-links

This patch improve patch
98ba39f250174783a32cad76842cf8f34a4281aa git-hooks:
  overwrite the windows-links not with linux-links

The main problem with the previous patch is that
when an alias is set for git, this alias will not
map in the Shell script passed, better you
use a git-symlink.

The other problem was the behavior of cygwin-bash
and win-git-bash that is not always the same, e.g.
   - winlnk=$(cmd /C ... 2>&1) it hang infinity,
      line 123
   - the 'ln' made not link, it only copy the files

Improving the FOR in ./git-hooks/README, when
you have the copied files, need a other del command

The patch can not set the windows-links, only output
the ./git-hooks/README
But in Win 11 should be possible, it is not needed
admin-rights for the 'mklink' command

Change-Id: Icecdb96e65fe2bba1270dfad2ac1af5af145925a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148462
Tested-by: Jenkins
Reviewed-by: Juergen Funk <juergen.funk_ml@cib.de>