sw: fix stale SwNumberPortion follow flag
Somehow a SwNumberPortion can be split into 2 SwNumberPortions and the
2nd one ends up with a '\t' character that is then removed from the
portion and via some HookChar thing converted to a SwTabLeftPortion.

To add further insult, the SwTabLeftPortion is inserted between the
first SwNumberPortion and the empty SwNumberPortion, and then
SwLineLayout::CalcLine() deletes the empty one but the first one still
has the m_bHasFollow flag set.

Change-Id: Ic902ed7f35d7faf168bef636b869e5cd9e2435ec
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148555
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>