tdf#140659 gtk a11y: Don't unset model when freezing icon view
As described in Change-Id I10249bbd8c684e89174ba91ce4690d37c24b5d5c
("tdf#153657 gtk3 a11y: Use IconView item tooltip as a11y desc"),
freezing the gtk3 icon view previously resulted in the accessible
child objects not being created at the time of insertion,
so setting the accessible description on them doesn't work.

As a result, e.g. the Orca screen reader would not properly
announce the focused item in Math's elements dock window when using
the gtk3 VCL plugin.

This was caused by clearing the model in
`GtkInstanceIconView::freeze` (and resetting it in
`GtkInstanceIconView::thaw`).

Stop doing that to fix the problem.

Thanks to Caolán for this suggestion!

This also allows to revert

    commit 9a08c471a08e6c21fb4e5303bf9a24cd55ae5e87
    Date:   Wed Dec 7 13:20:03 2022 +0000

        Resolves: tdf#152411 clear before freeze to let gtk a11y drop reference

in a follow-up commit since the problem described in
tdf#152411 also no longer happens with this in place.

Also switch the previous check + SAL_WARN to an assert
now that the underlying cause has been addressed.

Change-Id: Id0c241d68ec4fbf933312008f7d0ee86bd3eab0c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148535
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>