sw floattable: ignore height of masters in lcl_CalcMinRowHeight()
The problem was that the bugdoc has a split row and the minimum row
height was only enforced for the master row, not the follow one.

It seems this behavior started with commit
94ea3ba58867a4f786903d0c617654b0c5e204b6 (tdf#104425 sw: split rows
w/large min height (fix layout loop), 2016-12-08), because that had a
case where the minimum row height didn't fit a page, but moving to the
next page didn't help, so we hit a loop. The trouble is that when we do
have enough space in the body frame, then not enforcing a minimal row
height for each part of a split row leads to bad layout, since the row
height will be smaller, so content after the table moves up.

Fix this by checking if the body frame has enough space
and not consider the height of the master + previous follows in that
case.

This is just for floating tables at the moment, doing this in general
may need more work. With this, orig-nocompat.docx from tdf#61594 renders
correctly with SW_FORCE_FLY_SPLIT=1.

Change-Id: I922434db8455aba03ebeab9eae845003935f1d77
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148595
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins