tdf#153801 fix transparency in About dialog on gen/kf5
This reverts
    commit 8d3ded50fcf168d6b684091ba4683bd2bb7a2d30
    Author: Noel Grandin <noelgrandin@gmail.com>
    Date:   Sun Feb 5 11:29:27 2023 +0100
    make create_virtual_device not use a separate alpha layer

I made the change to make the
  SalInstanceWidget::create_virtual_device
and
  GtkInstanceWidget::create_virtual_device
the same, to make my life easier in landing the transparency->alpha
patch.
But that cannot work - the only reason the gtk path works is because gtk
uses combined alpha (ie. 32-bit) but that requires gtk doing some lying
and general fudging to the upper vcl layers because they don't know
about combined alpha.

Change-Id: If0e5046b05302a8f41b87846123685ed8f8e86f4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148597
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>