lok: capture and publish the modified status in save result
This is necessary to flag to the storage whether or not
the contents of the file being uploaded has any user
modifications or not. This is typically used to optimize
the versioning and storage utilization.

We also add the time when saving started and the duration
it took. The timestamp is to figure out if there has been
subsequent activity/commands that could have modified
the document. The duration is to help Online throttle
saving if too frequent.

Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>
Change-Id: Id6d8e629ef9b6e723d1d8b84d64fc7741b997bc5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149985
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>