tdf#139618 Add WordArt fill types in pptx import
This fix uses a similar approach as in WpsContext::onEndElement, only
that here moTextOutlineProperties and maFillProperties are still
available whereas in WpsContext the properties need to be collected
from CharInteropGrabBag.
I have adapted unit test testTdf125573_FontworkScaleX because now the
gradient is imported, see comment in former version of the test.

Change-Id: I989ee475867c97b9f93c7e65cd25483611eec7bf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148687
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>