tdf#112287 tdf#154129 writerfilter framePr: fix vAnchor default
This patch depends on previous patches for tdf#154129.

yAlign takes precedence over w:y. HOWEVER, there is a clause
in the documentation that complicates everything.

"If [yAlign is] specified, the position for this attribute
supersedes any value which is specified in the y attribute,
and that value is ignored, UNLESS the vAnchor is set to 'text',
in which case any relative positioning is not allowed,
and [yAlign] is itself ignored"

So if vAnchor is not defined, it defaults to text (FRAME),
but if yAlign is defined then it should mean margin
(or else yAlign has no meaning), unless of course w:y
is also defined, in which case "text" anchor has meaning again
and this unit test shows that w:y wins out in this case.

The anchor is not affected by wrapping.

[P.S. This was already set correctly in the code, but worth noting:
 "The standard states that the default value for the vAnchor attribute is page.
 Word uses a default value of 'text' for this attribute."]

make CppunitTest_sw_ooxmlexport5 CPPUNIT_TEST_NAME=testTdf112287
make CppunitTest_sw_ooxmlexport5 CPPUNIT_TEST_NAME=testTdf112287B

Change-Id: I953afa89e3c0974a16d43500ce590e1a481c2251
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148691
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins