tdf#154129 writerfilter framePr: check whole style inheritance: W/H
This patch depends on HAnchor, which created vProps.

This patch changes the logic a little bit from the previous code.
Before, the paragraph has to specify something other than 0 (auto)
in order to be acceptable. Well, I don't see any reason why
direct formatting shouldn't be able to override a non-auto style value,
so we'll make the change and see how it plays out.

Errata says:
"The standard states that 0 is a valid value of the h attribute.
 HOWEVER Word treats a value of 0 on this attribute as a value of auto."

and:
"The standard states that the values of the h attribute
 are defined by the ST_TwipsMeasure simple type.
 HOWEVER Word restricts the value of this attribute
 to be at least 1 and at most 31680."

There were no existing ooxmlexport unit tests where the old and new method
produced different values.

Change-Id: I48934598c1bca8ffc99b97b114d77505f3341105
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148715
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins