sc a11y: tdf#139322 Don't steal focus from "Manage Names" dialog
When opening the "Manage Names" dialog from the name box,
move setting the dialog to after calling `ReleaseFocus_Impl`,
since that one sets focus to the active view, and may thus
steal the focus from the dialog otherwise. (Observed
with gtk3 on Linux and on Windows, while focus was still
in the dialog with kf5 or gen.)

Setting the dialog afterwards ensures that the dialog
gets focus when it opens, while still ensuring that the
focus is no longer in the name box when the dialog closes
again.

Change-Id: I47f12d115a747ecda1c11c04282bd848eb398d9c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148775
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>