tdf#154129 writerfilter framePr: deduplicate w:h
This is not quite a No Functional Change commit.
In the previous code, a zero value for H was not added as a property
if there was no parent style.

However, it seems that pretty much only RTF
fails to typically have styles defined,
since even an unspecified "Normal" style is checked.
(fdo77890.docx also followed the "else" path
 because it basically had an empty styles.xml file.)

That means that in all but the most bizarre cases,
the "no style" case would never be reached anyway.
That really eliminates any worries about the deduplication going bad.

In any case, it makes it the same as width (and most DOCX files)
which had been doing this in the first place.

For consistency and code simplicity I'll go ahead with the change.

Change-Id: I65eed50ba83397284f21bb20ef3bbf10e02cce04
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148805
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>