tdf#154129 writerfilter framePr: deduplicate w:x
This is not quite a No Functional Change commit.
In the previous code, a zero value for X was not added as a property
if there was no parent style.

Basically only RTF format ever has no "parent style",
so this should be a rather safe change to make.

I've done the same thing for others as well.

Change-Id: Iaf4eb45b3adae88622fa9be16b5237db101a631e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148807
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>