sw floattable: reject small(er than min height) row master at page bottom
The problem was that the second row in the doc model was split between
page 1 & page 2, but it was expected that row 2 is entirely on page 2.

The underlying reason seems to be similar to what commit
bf8e59380a9a7517b73378741e7f275fa54ccdc0 (sw floattable: ignore height
of masters in lcl_CalcMinRowHeight(), 2023-03-10) fixed: again the
minimum row height is meant to be enforced for each row frame of the
row, so in case the remainings size on the page is 566 twips and the
minimum row height is 1134, then we should not split the row, instead we
should move the entire row to the next page.

Fix the problem by checking for this (relatively) at the start of
SwTabFrame::Split(), so we don't even try to split the row in this case.

I assume that this would be correct to do in general, but have no such
sample to confirm this in Word, so keep this specific to tables in split
flys for now.

Change-Id: Ic9171683ef7e14a70f7f8d5305198bbc16277d92
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148834
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins