sw floattable, legacy: restrict height when using bottom margin area
The bugdoc was 1 page in Writer, but 2 pages in Word. There the entire
3rd row went to the next page.

The problem seems to be that the legacy mode in Word seems to allow
going to the bottom margin only in case the height of the floating table
still fits the body frame height, while we simply allowed the usage of
the bottom margin area unconditionally.

Fix this by extending GetFlyAnchorBottom() to explicitly check for the
body frame height.

Note that this doesn't solve cases where the current height fits the
body frame, but the new bottom won't; that still needs fixing in a
follow-up commit.

Change-Id: Ic55b93e7b53fc45a837c3245c0091f2c93d028aa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148944
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins