tdf#154071 UI rename gradient back to 'Rectangular'...
Commit cff737a7a3caaa88cc524062feb78fd42295b077 had wrongly changed
'Rectangular' to 'Square' (API RECT, ODF rectangle) and 'Square' to
'Quadratic' (API SQUARE, ODF square).
This patch reverts 'Square' to 'Rectangular' and 'Quadratic' to
'Square (Quadratic)'. It uses 'Square (Quadratic)' instead of simple
'Square' to help users to find the correct type.
In addition, use of the wrong type names in comments is fixed.

Change-Id: Ibbd90be2cceb65c9857e473cb74c39df1b9530f1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149004
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>