sc drawstyles: Delay the destruction of SdrItemPool
Nowadays SfxItemPool_Impl::mpSecondary is ref counted, so
there is no fear in destroying the draw layer before the doc
pool.

This kind of change is needed for drawing styles, as their
item set references the SdrItemPool with a raw pointer, but
they are destroyed as part of the styles pool, much later
than the draw layer.

Change-Id: I169bbacfbdbec2121217b13293b6f495898fe2e4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149090
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>